requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.39.0

openai==1.69.0
langchain-core==0.3.49
langchain-openai==0.3.11
langchain-ollama==0.2.1
langgraph==0.3.21
sqlalchemy==2.0.23
psycopg[binary]
